Create Download a CSV file of all users in a specific AD group, including all nested groups.  Show user names.
Replace "YourGroupName" with the name of the specific AD group you want to query.

The script defines a Get-UsersFromGroup function that recursively retrieves users from groups, including nested groups. It checks the objectClass of each member and adds users to the $userInfo array.

Users' names are then exported to a CSV file named "UserList.csv" in the current directory using Export-CSV with the -NoTypeInformation parameter.

Running this script will generate a CSV file containing the names of all users in the specified group and its nested groups.